+++ native/window/src/index.d.ts
@@ -1,4 +1,6 @@
 export interface NativeWindowApi {
+    setGlFramebufferTransparent(nativeData: Uint8Array, transparent: boolean): void;
+    setTransparent(nativeData: Uint8Array, transparent: boolean): void;
     create(nativeData: Uint8Array, onFrame: () => void, onState: (active: boolean) => void): ModalFrameController;
 }